@@ -1,11 +1,11 @@
-﻿using Microsoft.AspNetCore.Hosting;
+﻿using Microsoft.Extensions.Hosting;
 using Nm.Lib.Host.Web;
 
 namespace Nm.WebHost
 {
     public class Startup : StartupAbstract
     {
-        public Startup(IHostingEnvironment env) : base(env)
+        public Startup(IHostEnvironment env) : base(env)
         {
         }
     }
